+++ changelog.md
@@ -8,6 +8,7 @@
 - Bugfix: Group membership watch was not re-established
 - Bugfix: Wrong node state was displayed in the GUI
 - Bugfix: Fixed handling of failed futures in the benchmark
+- Bugfix: Fixed calculation of in flight calls for a cluster
 
 ### Version 0.2.0 (Beta-2) - 28.11.2016
 - New Feature: Java 8 is now required to build the project